Command Buffer Variant
A variation on the effect shown in the video. Instead of using a second
camera, a Command Buffer is set up to render the geometry it needs to
just before Image Effects are applied.

This removes the need for Camera.SetReplacementShader and is an example
of how you can better control what object do or do not emit a glow.

Also switched some binary assets to text serialization.